修改U1,U2...U8的Java类型代表，并为其添加注释

+++ src/main/java/buddha/jvmabi/attribute/CodeAttr.java
@@ -18,8 +18,8 @@ import java.io.Serializable;
 public final class CodeAttr implements Serializable, IJvmAttribute {
 
     private int length;
-    private short maxStack;
-    private short maxLocals;
+    private int maxStack;
+    private int maxLocals;
     private byte[] code;
     private CodeExceptionInfo[] exceptionTable;
     private IJvmAttribute[] attributes;
@@ -35,7 +35,7 @@ public final class CodeAttr implements Serializable, IJvmAttribute {
         maxStack = reader.readU2();
         maxLocals = reader.readU2();
         code = new byte[reader.readU4()];
-        for (int i = 0; i < code.length; i++) { code[i] = reader.readU1(); }
+        for (int i = 0; i < code.length; i++) { code[i] = (byte)reader.readU1(); }
         exceptionTable = new CodeExceptionInfo[reader.readU2()];
         for (int i = 0; i < exceptionTable.length; i++) {
             exceptionTable[i] = new CodeExceptionInfo(reader.readU2(),
